Mapper chyba działa - wprowadzono kilka poprawek literówek itd

+++ src/main/java/org/example/api/le_systeme_solarie/SolarSystemPlanetDetailsResponse.java
@@ -19,9 +19,9 @@ public class SolarSystemPlanetDetailsResponse {
     private Double meanRadius;
 
     //TODO: Tu ewentualnie można dodać jakieś nowe informacje i zastanowić się dlaczego discoveredBy - nie działa...
-    public Integer getNumericId() {
+    public Long getNumericId() {
         try {
-            return Integer.parseInt(id);
+            return Long.parseLong(id);
         } catch (NumberFormatException e) {
             return null;
         }
